Removed fixed redshift-path

+++ redshift-tray/redshift.cs
@@ -24,7 +24,6 @@ namespace redshift_tray
 {
   public class Redshift
   {
-    private readonly static string REDSHIFTPATH = Path.Combine(System.AppDomain.CurrentDomain.BaseDirectory, "redshift.exe");
     public readonly static int[] MIN_REDSHIFT_VERSION = { 1, 10 };
 
     private Process RedshiftProcess;
@@ -39,17 +38,17 @@ namespace redshift_tray
       }
     }
 
-    public static RedshiftError Check()
+    public static RedshiftError Check(string path)
     {
       Main.WriteLogMessage("Checking redshift executable", DebugConsole.LogType.Info);
 
-      if(!File.Exists(REDSHIFTPATH))
+      if(!File.Exists(path))
       {
         Main.WriteLogMessage("Redshift executable not found", DebugConsole.LogType.Error);
         return RedshiftError.NotFound;
       }
 
-      string[] version = StartAndWaitForOutput("-V").Split(' ');
+      string[] version = StartAndWaitForOutput(path, "-V").Split(' ');
 
       if(version.Length < 2 || version[0] != "redshift")
       {
@@ -85,33 +84,36 @@ namespace redshift_tray
       return (majorversion == MIN_REDSHIFT_VERSION[0] && minorVersion >= MIN_REDSHIFT_VERSION[1]);
     }
 
-    public static Redshift Start(params string[] Args)
+    public static Redshift StartContinuous(string path, params string[] Args)
     {
+      if(Check(path) != RedshiftError.Ok)
+        throw new Exception("Invalid redshift start.");
+
       if(Instance != null && !Instance.RedshiftProcess.HasExited)
       {
         Instance.RedshiftProcess.Kill();
       }
 
-      Instance = new Redshift(Args);
+      Instance = new Redshift(path, Args);
       return Instance;
     }
 
-    public static string StartAndWaitForOutput(params string[] Args)
+    public static string StartAndWaitForOutput(string path, params string[] Args)
     {
-      Redshift redshift = new Redshift(Args);
+      Redshift redshift = new Redshift(path, Args);
       redshift.RedshiftProcess.WaitForExit();
 
       return redshift.GetOutput();
     }
 
-    private Redshift(params string[] Args)
+    private Redshift(string path, params string[] Args)
     {
       string arglist = string.Join(" ", Args);
 
       Main.WriteLogMessage(string.Format("Starting redshift with args '{0}'", arglist), DebugConsole.LogType.Info);
 
       RedshiftProcess = new Process();
-      RedshiftProcess.StartInfo.FileName = REDSHIFTPATH;
+      RedshiftProcess.StartInfo.FileName = path;
       RedshiftProcess.StartInfo.Arguments = arglist;
       RedshiftProcess.StartInfo.UseShellExecute = false;
       RedshiftProcess.StartInfo.CreateNoWindow = true;
